Replace usage of jdk logging with jboss-logging
The sample project is currently using jdk logging and tied to it.
The jboss-logging is an Hibernate-default logging framework that
supports all main stream implementations. We shall follow Hibernate
convention by switching to jboss-logging. Now we can change logging
implementation at deployment time.

For the implementation, we choose logback in pom.xml.